Change AnsibleVars to string in CRD schema
AnsibleVars can't be a map in the CRD schema because the map values
could be any other types: another map, a string, list, etc. Using
map[string]interface{} is not possible due to controller-gen validation,
which fails with:

not a supported map value type: *ast.InterfaceType

Instead, use string as the type, and handle unmarshaling the string into
YAML in the code. Once unmarsahled into YAML, each top level key is set
as an ansible host var in the generated inventory.

Signed-off-by: James Slagle <jslagle@redhat.com>