Rework AnsibleExecution to use CreateOrPatch
The AnsibleExecution is reworked to use controllerutil.CreateOrPatch, so
that multiple AnsibleEE CR's are not inadvertently created. Labels are
also added so that the AnsibleEE CR's can be uniquely identified per
node or role when listing them.

controllerutils.SetControllerReference is called to ensure that the
AnsibleEE CR's are deleted when the DataplaneNode/DataplaneRole are
deleted.

Signed-off-by: James Slagle <jslagle@redhat.com>