Add Deploy boolean and refactor deployment calls
Adds a new boolean to the DataPlaneNode and DataPlaneRole CRD's called
Deploy. Deploy controls whether any of the ansible is triggered when the
node or role is reconciled.

The deployment logic is also refactored out into a single deployment.go
file within the deployment package, and the package is moved directly
under pkg/ instead of the not needed dataplane/ directory.

Signed-off-by: James Slagle <jslagle@redhat.com>