Add status and condition coordination with ConfigureNetwork
The logic on starting the ConfigureNetwork AnsibleEE is now coordinated
with conditions, and guarded with a ConfigureNetworkReadyCondition.

The condition handling is summarized as:
 - when initially empty, we initialize the condition to unknown and
   requeue.
 - when unkown, we start the ConfigureNetwork AnsiblEE, set the
   condition to False, and requeue
 - when False, we check the status of the triggered AnsibleEE job, if
   the job status is succeeded, we set the condition to true. If the job
   status is not succeeded, we requeue.
 - when True, we assume ConfigureNetwork is already completed.

The GetAnsibleExecution and GetAnsibleExecutionJob functions are added
to look up existing executions and jobs.

A top level Deployed boolean is added to the status, to indicate the
overall deployment state.

Signed-off-by: James Slagle <jslagle@redhat.com>