Add AnsibleTags to CRD
Add support for ansible tags during the deployment. As there will be
additional fields to control the deployment, a DeployStrategy Section is
added as well. Within DeployStrategy, the Deploy bool is moved, and the
AnsibleTags field is added. Additional fields controlling how the
deployment is run can be added here as well.

At the end of a successful deploy, instance.Spec.DeployStrategy.Deploy
is always set to false. This is because we do not want additional
reconcile requests to trigger a new deployment unless the user has
explicitly set the Deploy bool to true (or explicitly done by another
operator).

Signed-off-by: James Slagle <jslagle@redhat.com>