Make node's role field optional
If an OpenStackDataPlaneNode CR is created before any
OpenStackDataPlaneRole exists then do not fail with
the error below.

This patch assumes we want to support creating nodes
without roles. If we want to refuse to create nodes
unless there is a role, then a validation should be
added to the API so that the role field of the node
is required.

{
  "controller": "openstackdataplanenode",
  "controllerGroup": "dataplane.openstack.org",
  "controllerKind": "OpenStackDataPlaneNode",
  "OpenStackDataPlaneNode": {
    "name": "network-edpm-compute-0",
    "namespace": "openstack"
  },
  "namespace": "openstack",
  "name": "network-edpm-compute-0",
  "reconcileID": "ada7594e-65b2-446e-b4bc-49d8989d5ad3",
  "error": "OpenStackDataPlaneRole.dataplane.openstack.org \"\" not found"
}

Signed-off-by: John Fulton <fulton@redhat.com>