Add NovaExternalCompute to deploy
To deploy nova-compute services on the EDPM nodes, a NovaExternalCompute
CR is created as the last step of the deployment. The
NovaExternalCompute is labeled with:

openstackdataplanenode=<uuid>

where <uuid> is the UUID of the openstackdataplanenode instance.

The ReadyCondition is checked on an existing NovaExternalCompute CR to
check for the deployment status.

Signed-off-by: James Slagle <jslagle@redhat.com>