Group nodes on roles with only the OpenStackDataPlaneNode.Spec.Role field
Removes the list of nodes in a role that was set with the
OpenStackDataPlaneRole.Spec.DataPlaneNodes field. Instead, the node and
role grouping association will only be captured in the
OpenStackDataPlaneNode.Spec.Role field.

This simplifies the grouping logic and ensures that the single source of
truth is only kept in one place. It removes the need to keep the list of
nodes in sync on the role. Nodes are also no longer reconciled by a Role
reconcile reequest. Instead, nodes can be created in bulk using the
OpenStackDataPlaneNodeList type.

To simplify querying or filtering for nodes in a role, the
"openstackdataplanerole" label is added to nodes that are in a role. The
label's value is the role name.

This patch also updates the README.md with this change, and adds
additional docs to describe the behavior.

Signed-off-by: James Slagle <jslagle@redhat.com>