Add OpenStackDataPlaneRole deploy
Adds the deploy logic to OpenStackDataPlaneRole. The DeployStrategy
struct is added to the role, and is the same struct used on the node.

OpenStackAnsibleEERunnerImage and NetworkAttachments are also added to
the role spec. These properties must be set the same during a role
deploy, so they are top level on the role. As all nodes in the role are
in the inventory passed to the AnsibleEE, these values must be the same
during a role deploy.

Signed-off-by: James Slagle <jslagle@redhat.com>