Configure OpenStackDataPlaneNode as Ceph clients
If an OpenStackDataPlaneNode has any extraMounts items
where extraVolType is Ceph, then call the Ansible Role
edpm_ceph_client_files and pass it the mounted files.
This results in nodes having Ceph configuration files
(e.g. ceph.conf or a cephx key) so that the OpenStack
services on those nodes can be clients of a Ceph cluster.

Closes: OSP-22956

Signed-off-by: John Fulton <fulton@redhat.com>