Add DeployNovaExternalCompute to role deploy
As the last step in the role deploy, DeployNovaExternalCompute is called
to deploy the nova services. This commit refactors the function since
the returns need to be handled differently whether it's a node deploy or
a role deploy.

During a node deploy, only the single created (or patched)
NovaExternalCompute needs to be checked for errors or the
ReadyCondition. However, during a role deploy, each NovaExternalCompute
for each corresponding node needs to be checked. If any of the
NovaExternalCompute CR's is in error, or not ready, then the entire role
is set as in error or not ready.

The call to DeployNovaExternalCompute is refactored out of the Deploy
func and called directly from the node and role controller instead so
that the return handling can be different.

This change also switches to use the standard ReadyCondition from
lib-common instead of the custom types DataPlaneNodeReadyCondition and
DataPlaneRoleReadyCondition. The custom types are not needed as a custom
message can already be set on the standard ReadyCondition. The standard
ReadyCondition is also used to reflect the overall status in the CLI and
UI.

Signed-off-by: James Slagle <jslagle@redhat.com>