Handle when ansibleVars is missing from node or role
If ansibleVars was not set in the role, then use the
ansibleVars from the node and vice versa. Otherwise
merge them. If ansibleVars is not in either the role
or node, then the inventory is created without them.
This is OK since ansibleVars is for extra vars.

Follow fix to 5295e54f0bf27624a4549d9e7643c28f100ed8b2

Related: OSP-23184

Signed-off-by: John Fulton <fulton@redhat.com>