Trigger role reconcile for NovaExternalCompute update
OpenStackDataPlaneRole was not being reconciled when the associated
NovaExternalCompute's it had started were updated. This was due to 2
reasons.

First, the manager was not setup to Own the resources. This patch also
adds the other missing Owns, and adds the kubebuilder rules for the
other resources that both OpenStackDataPlaneRole and
OpenStackDataPlaneNode should own.

Second, the controller reference set on NovaExternalCompute was for the
node. In the case of a role deploy, this needs to be the role, so that
the role reconcile is triggered. The function signature of
DeployNovaExternalCompute needed to change to take an additional owner
parameter so that the owner can be set as either the node or role
depending on which type of deploy is triggered.

With these changes, roles are now reconciled a final time after all
NovaExternalCompute's are Ready, which is needed to set the final ready
state to true on the role.

Signed-off-by: James Slagle <jslagle@redhat.com>