Drop explicit requeues for OpenStackAnsibleEE
Explicit requeues (ctrl.Result{requeue:True}) are not strictly needed.
They were added before we had the DataPlane Manager properly watching
the OpenStackAnsibleEE resources with the owner reference set. Now that
we have that in place, we do not need these explicit requeues. As the
status of OpenStackAnsibleEE resources change, the dataplane controllers
will be automatically notified and a new reconciliation loop will be
triggered.

Signed-off-by: James Slagle <jslagle@redhat.com>