Add Composable Service CRD and controllers
Adds a new top level CRD, OpenStackDataPlaneService which represents an
ansible based service that can be selectively included in a dataplane
role deployment.

The structure of the OpenStackDataPlaneService.Spec.Role field matches
that of the OpenStackAnsbileEE.Spec.Role. This will allow for moving the
hardcoded ansible roles and tasks out of the code and into a CR of the
OpenStackDataPlaneService kind.

Services are added to role with the Services field on
OpenStackDataPlaneRole.

Signed-off-by: James Slagle <jslagle@redhat.com>