Allow nova configurability on the EDPM node
This patch exposes a nova specific template struct within the
DataplaneRole and Node that allows the end user to pass compute
specific configuration to the EDPM node. The template in the Node has
precedence over the template in the Role. If neither is defined then it
indicates that the EDPM node does not need compute services to be
deployed (i.e. it is not a compute node)