Use DeploymentReadyCondition and separate from ReadyCondition
Separates the deploy readiness condition handling out of ReadyCondition
by using the DeploymentReadyCondition from lib-common.

ReadyCondition was previously overloaded. It was the only condition
checked from the resource's IsReady functions, and then the condition
was set based on the result of IsReady. Instead IsReady will check the
delpoyed flag and set ReadyCondition as a result.

Signed-off-by: James Slagle <jslagle@redhat.com>