Add SetupReadyCondition
SetupReadyCondition tracks the status of the resource setup separate
from the deployment. Setup includes creating other child resources such
as the ConfigMaps and Secret verification, as well as synchronizing
labels. Setup effectively indicates all the reconcile logic outside of
the actualy deployment.

In order to properly initialize the SetupReadyCondition, this also moves
the condition initialization to be earlier in the reconcile function.

The "defer func" is also moved to the top of the reconcile function for
clarity.

Signed-off-by: James Slagle <jslagle@redhat.com>