Add kuttl test for deployment
Adds a step to the dataplane-deploy-no-nodes kuttl test to toggle the
deploy flag on OpenStackDataPlane. Since there are no nodes, the rest
runs relatively quickly (< 3 minutes) and is a good smoke test for the
deployment logic. The state of the created OpenStackAnsibleEE resources
are validated in the steps assert (01-assert.yaml).

This commit also fixes bug in the condition setting for the
OpenStackDataPlane resource that was found by the test.

Signed-off-by: James Slagle <jslagle@redhat.com>