Selective BaremetalSetTemplate
Only copy roleSpec.BaremetalSetTemplate if we have BaremetalHosts. This
change moves the role.Spec.BaremetalSetTemplate to ensure it's only
copied when baremetal nodes exist. Else we need to define the required
paramaters of this Spec during deployment.

Since it can be ommitted in the DataplaneNodeSpec [1], we should respect
that in the Controller code and not set it unless we expect to have
baremetal managed nodes included in the deployment [2].

1: https://github.com/openstack-k8s-operators/dataplane-operator/blob/main/api/v1beta1/openstackdataplanerole_types.go#L40
2: https://github.com/openstack-k8s-operators/dataplane-operator/blob/main/controllers/openstackdataplane_controller.go#L250