Add initial envtest for OpenStackDataPlane
This patch mainly just sets up the envtest environment so that it's
ready for Ginkgo tests to be defined in future commits. For now, we just
validate that we can deploy OpenStackDataPlane and
OpenStackDataPlaneRole Kind objects and that they have Spec and Status fields
registered. Creating a OpenStackDataPlane object should also create a
OpenStackDataPlaneRole, so this is also asserted and checked in these
initial tests.