Migrate ValidateNetwork to validate-network composable service
Since ValidateNetwork was the last service from
pkg/deployment/network.go, that file can be removed.

Also renames the configure-network service to use an underscore in the
service file name to match conventions.

This also changes the InitConditions functions for the node and role to
initialize conditions for the composable services.

Signed-off-by: James Slagle <jslagle@redhat.com>