Remove openstackdataplaneservice controller
We're not doing anything in the openstackdataplaneservice
reconciliation. The controller is kind of redundant atm. I
suspect this would be some kind of static configuration of
the play to run. We could have used a configmap instead, but
I guess CRD will give us the opportunity to vlaidate/default
the spec attributes.

This patch proposes to remove the controller.