[envtest]Enhance test make target
I switched to ginkgo cli to execute the env test. This allows better
configurability of the execution. To still run the existing unit test
there is a separate go test call and a ginkgo call in make test target.

Added GINKGO_ARGS so that we can pass parameters to the test runner,
e.g.
  make test GINKGO_ARGS="-v --focus 'Should have created a OpenStackDataplaneRole'"
will run a single test case with verbose logging.

Also added parallel test execution and for that disabled the metrics
service as that wants to bind to the same port (8080) from each parallel
tests.