Move OpenStackAnsibleEEImage field to OpenStackDataPlaneService
The image used by ansible-runner needs to be service specific. The
default value remains the same, and that value will be used when an
image is not specified.

For custom services enabled with the composable service interface that
may need custom ansible content, a different value can be specified for
the image.

Adds a kuttl test with a custom service using a custom image. The tests
expects an OpenStackAnsibleEE resource that never succeeds since the
custom image does not actually exist.

Also moves setting the SetupReadyCondition to true earlier in the
reconcile functions. The condition should be set true as soon as all the
setup related tasks are done, which can happen before the deployment is
triggered. For cases where the resources are created with
DeployStrategy.Deploy=true, the condition will now be marked true before
the deployment is started as expected.

Signed-off-by: James Slagle <jslagle@redhat.com>