Add deployProtection annotation to osdpnode
When a node is already deployed, we need to protect against eroneous
Ansible executions. This is essential to ensure predictability in our
deployments and ensure nodes are only being touched when there is
explicit intent to do so.

This change annotates the OpenStackDataPlane objects at the end of the
deployment. This annotation will need to be explicitly removed to
trigger a redeploy.

Add methods to check managed node annotations

This change adds methods to remove the redeployProtection annotation
from controller managed objects

Signed-off-by: Brendan Shephard <bshephar@redhat.com>