Use AEE instead of the Job to set status
This change uses the AnsibleEE object rather than the job to determine
the status of the deployment. This ensures that we can still set the
status even if the jobs are cleaned up after execution and prevents us
from being stuck in an unknown state.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>

Add tidy Makefile target to pre-commit

This change ensures that pre-commit executes go mod tidy for both the
root directory and also the api/.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>