Wait for DNSMasq to be ready and reconcile for DNSMasq changes
If dnsmasq service exists and not ready, we should wait for it.
Also, if there are any changes in DNSMasq we should reconcile.
This won't normally happen as the service would be created by
the controlplane and won't be changed.

But it's posible that DNSMasq service has been deleted and role
has to be reconciled once it's reconciled.