Remove not required conditions
When we are using pre-provisioned nodes, we don't need the services
provided for baremetal nodes such as:

- RoleBareMetalProvision
- RoleDNSData
- RoleIPReservation

This change ensures that these conditions are removed from the status
to avoid causing any undue confusion and/or concern for our users.
Additionally, in situations where the CR is created with NO
baremetalSetTemplate.BaremetalHosts. Then we don't bother initialising
these status.conditions to begin with.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>