Only initialise services if we're deploying
If a CR is created with deploy = false, we don't need to initialise the
service statuses at this point. This change adds a conditional to only
initialise service conditions if a) we have services in the list, and b)
we are going to be trying to deploy them base on
spec.deployStrategy.deploy being true.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>