Update nic-config tests, samples and docs
This change updates the nic-config implementations, samples and
documentation to reflect the changes made in:
https://github.com/openstack-k8s-operators/edpm-ansible/pull/243

This allows us to remove the additional mounts for the AnsibleEE pods
and instead allows the user to define their network configuration via
the Ansible variable directly.

edpm_network_config_template takes a string variable via the CRD
definition.

Depends-On: https://github.com/openstack-k8s-operators/edpm-ansible/pull/225
Signed-off-by: Brendan Shephard <bshephar@redhat.com>