Return error instead of nil _err
In the node and role controllers, we currently return _err which at this
point has only been initialised as the named return value in the
function signature. Accordingly, this will always be nil at the point we
are returning. This change returns the err from the previously called
function before setting _err to the same and returning from the
function.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>