Remove -leader-elect for controller manager
We use a single replica for the controller-manager and there is
probably no need to use multiple of them for high availability
of the operator pods. This patch removes the leader election
with single replica. When we move to use multiple replicas
we can use ENV var to patch it.