Fix ExtraMounts
ExtraMounts were being dropped and not added to the created
OpenStackAnsibleEE resources. The reason was because as each composable
service is deployed, aeeSpec.ExtraMounts is reset back to the original
value so that there is no cross over between mounts from individual
composable services. The reset was not working because the initialied
value of aeeSpec.ExtraMounts did not have a lenght set, causing the
copy() function to not actually copy any data.

The commit switches to using make(...) with a length parameter to ensure
the copy() works as intended.

Adds a new kuttl test, dataplane-extramounts, to test this specific
functionality.

Signed-off-by: James Slagle <jslagle@redhat.com>