[DataplaneService/nova]Add migration ssh key
The nova role in edpm-ansible requires[1] an ssh key-pair that is then used
as to authenticate the nova user when it copies data between compute
nodes during migration.

A placeholder Secret name for this key-pair is now added to the nova
DataplaneService definition. The deployer (human, ci_framework, or
install_yamls) needs to make sure that a secret is created with an SSH
key-pair in it.

[1] https://github.com/openstack-k8s-operators/edpm-ansible/pull/446