Add image_prefix ansible vars in dataplane crs
In downstream, the osp18 image name starts with
rhosp18-openstack and upstream one starts with openstack.

If we update the registry and image namespace in dataplane cr,
it will still pull the upstream image from defaults because
downstream image names are different.

This pr introduces image_prefix var to set image prefix and
avoid such issue.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>