Add mechanism to detect relevant spec changes
This change adds a Status field called ConfigChanged. We update this
field in cases where the current config differs from the previously
applied config. This is specifically used to detect changes between
spec.nodes and spec.nodeTemplate. If either of those have a variance,
then the ConfigChanged status field will be set to true.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>