Allow re-execution of AnsibleEE tasks
This change allows the re-execution of AnsibleEE jobs based on the
configHash generated from the OpenStackDataPlaneNodeSet CR inputs.

We do this by writing an Env var to the AnsibleEE CR and comparing that
with the latest hash from the NodeSet. If we detect a difference, then
we will delete and re-create the AnsibleEE CR's.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>