Drop container image ansible vars from samples
These image vars are not needed in the samples. There are defaults
already set in the corresponding roles in edpm-ansible.

Additionally, values set in the samples will override the
$RELATED_IMAGES* variables when those samples are used and customized by
users. Instead, we need the $RELATED_IMAGES* variables to take effect in
most cases so that users are using the corresponding images for the
right release and registry.

Signed-off-by: James Slagle <jslagle@redhat.com>
(cherry picked from commit eceb261442547339b9e3f6df4e9e68fe8e95f13c)