Use struct for deployment
Rather than passing copies of variables and structs to multiple
different functions throughout the deployment process. This change
defines a struct called Deployer. This struct is initialized with the
data it needs throughout the deployment process and independent
functions are switched to methods of this struct.

This reduces the required length and complexity of function signatures.
Objectively making it easier to follow the flow of code throughout the
deployment process.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>