Simplify signature of AnsibleExecution
The current function definition takes multiple components of the service
type resource. This change updates the function definition to instead
take a pointer to the OpenStackDataPlaneService where we can then
reference the various components. This simplifies the function signature
and makes it easier to keep track of where code is coming from.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>