Don't allow update/patch of deployment
Updating/patching of OpenStackDataplaneDeployment resource does
not have any effect when it'a already deployed. Also if it's
updated during the deployment it can lead to issues. The only
way at present is to delete the existing deployment resource
and create a new one.

Also, cleans up code around condition handling.