Remove cert-default name
Removed the line that basically makes multiple services that are
using the standard config use the same underlying cert.  I'm not
sure why I thought that was a good idea at the time, but it would
ultimately be less secure and adds unneccesary complication when
owner references will be added.

With this line removed, all services requesting a cert will have
a unique secret created containing the cert at
cert-<service>-<nodename>.