Add nic-config samples folder
This change copies some of the examples from EDPM Ansible into the
dataplane-operator samples. This allows users to better reference the
examples that can be in-lined in the NodeSet without having to jump back
and forward between edpm-ansible and dataplane-operator. This should
reduce confusion and improve user experience while trying to find
examples for what they are trying to do.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>