Add ContainerImages to NodeSet Spec and Status
Adds stronly typed fields for the dataplane ansible container images to
the NodeSet Spec and Status.

A defaulting webhook sets the fields to their default values from the
$RELATED_IMAGE* environment variables, or the upstream values if the
environment variables are not set.

The fields will be used by the OpenStackVersion controller to set
updated container image values during an OpenStack update.

When a deployment is completed, the values of the fields in the Spec are
saved to the Status so that the OpenStackVersion controller knows what
versions were used and if an update is needed.

Signed-off-by: James Slagle <jslagle@redhat.com>