Hash the entire NodeSet spec
This change removes the selective fields we were using to create
the ConfigHash. Initially, this was implemented to ensure we only captured user
provided values from the Spec, while removing those that are populated during
reconcile by the controller. For example, the values in the baremetalSetTemplate.
However, since we are now including the baremetalSetTemplate in the current hash,
there is no need to be selective about the fields we include.

This change removes the selective spec fields and instead hashes the entire
NodeSet spec section.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>