Temporarily fix issuers
The code currently sets the default issuer to
osp-rootca-issuer-internal, which is the label for the issuer - rather
than the name of the issuer itself (rootca-internal).

The proper fix for this is to use the label and then call something like
GetIssuerByLabel() to get the right issuer - while handling the error if
the issuer is not yet available.  We plan to add such a function to
lib-common by early next week - and will add code to use it then.

Until then, we can unblock folks trying to test out tls-e by setting the
issuer name  correctly.