Create a ServiceAccount for use by edpm nodes
Adds creating a ServiceAccount to the NodeSet reconciliation, with a
name equal to the NodeSet. The ClusterRole registry-viewer is set in a
RoleBinding (also named the same as the NodeSet) for the ServiceAccount.

The ServiceAccount will be used by EDPM nodes to podman login to the OCP
default internal registry to allow for pulling images. This allows for
mirroring images from their source (quay.io, etc) to the OCP internal
registry.

Since the ServiceAccount is named the same as the NodeSet, an additional
ansible variable is added to the inventory as edpm_nodeset_name. The
login tasks will use the value of this variable as the username.

The OpenStackAnsibleEE.Spec.ServiceAccountName field is set on all
OpenStackAnsibleEE's for the NodeSet so that the ServiceAccount
credentials are injected into the ansible pods. Ansible content will
then take over running the podman login command with the ServiceAccount
credentials on the EDPM nodes.

Signed-off-by: James Slagle <jslagle@redhat.com>