Fix cert names
The code in lib-common already adds "cert-" to the name of the secret
for the cert.  With our code, we end up with secrets with the name
"cert-cert-foo" which is redundant and at the least inconsistent with
the control plane.

Also, it looks like the check for the existence of the secret cert would
always fail (as it was looking for the wrong secret name), so we fixed it.

Finally, just modified the kuttl test to create a TLS combined-ca-bundle
more consistent with what we expect from the
openstack-operator/lib-common.